refactor(ofm_core/contracts): extract shared leaf-helper cluster
Move the 14 cross-cluster leaf helpers out of contracts/mod.rs into
contracts/helpers.rs: owned-player lookups (owned_player, owned_player_index,
contract_owner_team_id), wage/tenure maths (expected_wage, reference_player_wage,
importance_wage_multiplier, expected_contract_years, minimum_acceptable_wage,
is_insulting_wage_offer, round_up_to_nearest_thousand), the contract-date arithmetic
(player_age_on, remaining_contract_days, contract_days_remaining) and
backend_text_with_param.

Byte-faithful move: bodies extracted verbatim via script; private helpers promoted
to pub(crate) so the concern clusters and sibling ofm_core modules reach them via the
re-export (the already-pub(crate) expected_wage/expected_contract_years/
round_up_to_nearest_thousand keep their visibility). helpers.rs reaches the shared
types via `use super::*`.

387 ofm_core lib + 28 contracts integration tests pass; clippy clean in contracts/.
